requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
jinja2==3.1.4
google-api-python-client==2.147.0
google-auth==2.35.0
